requirements.txt
fastapi
uvicorn[standard]
passlib[bcrypt]
python-jose[cryptography]
email-validator
bcrypt==4.0.1
requests
